requirements.txt
click>=8.1.3
databricks_cli>=0.17.7
PyYAML>=6.0
rich>=12.5.1
setuptools>=49.2.1
pytest
wheel==0.41.0
pytest-html
pytest-click
flake8
black==23.7.0
pycodestyle==2.10.0
google-cloud-storage==2.10.0
pydantic==2.1.1
pydantic-yaml==1.1.1
google-cloud-secret-manager==2.16.2 # similiar to what we have in execution service.
boto3==1.28.12
cryptography==35.0.0 # latest version conflicts with dbt
requests~=2.31.0